move file to helpers folder and update import
